Refactor + started on subscribing
At the start we would hardcode the fact if a device is a owner of a puzzle and if another device is the subscriber of that puzzle based on the device name. We now want to change who is subscribed to what when we start a new game. So when we start a new game it will be possible to arrange the puzzles in a different order, changing the owner/subscriber(s) relation. If it's hardcoded this is not possible. Now a device only has a type and a puzzle asociated with it. A device can then subscribe to other puzzles to receive events.

Other changes:
- tests to reflect the changes made above
- SubscribePuzzleCommand started
- New exceptions
- Base test class
- New dependency so I can have tuples

This is one big commit becauce I am leaving at school and don't want to lose progress. Some work in this commit is needed for #11