Change architecture
- Compare with clean architecture
	- Core is dependenant on the Spring framework, but minimal usage of spring in domain and service (was core in clean)
	- Notification can make use of a websession instead of a generic class (tight coupling?)
	- Dependency injection is cleaner
	- Less verbose project (amount of clicks till you get somewhere)

If you truly ever want to rip spring out of the project even if you are using clean arch you will have to do major rewrites mostly in the web folder (clean) web package (layered). The service layer with the domain layer can be seen as the core and have minimal deps on spring (only for DI). So I think this way is fine as I'm not fighting as hard against the framework.

I think having clear boundaries in your code (no persistence logic in a service layer for example, no business logic in controllers etc) is the most important thing of clean arch, and the fact that you don't want your business logic to depend on a framework. We only use Spring DI in our service layer for example. So that's fine imo. We still use repositories and can stub them in the service layer so testing is easy and the service layer is not coupled to a database for example wich would be bad.

But we will see just trying things out :)

So differences clean => layered:
- core => domain
- core => service
- infrastructure/notification => service
- infrastructure/persistence => persistence
- web => web